Protect route
To prevent routes access without authentication every protected route and every single of its child routes must be provided with its own loader which calls the auth-checking requireAuth function.
To be able to redirect to protected route the requireAuth function appends that protected route pathname (and an optional message) as a query string to the authentication page url.
Finally, to prevent the user to navigate back to the authentication page after logging in the Form component is provided with the "replace" props.